Better handle changes to auth.provider config setting (#120)
After a site adminstrator changes the auth.provider setting from LOCAL
to KEYCLOAK, make Keycloak accounts for pre-existing users whenever
they enter their password, and keep track of their IDs on the Keycloak
server in UserMetadata.

Fix the reset password and change email flows to not create local
passwords or change emails for users with Keycloak accounts if the
Keycloak server is currently not configured, because this would cause
the local account and Keycloak to contain different information if the
Keycloak server is then re-configured (as might happen when a site
admin is learning Keycloak administration).  Add error messages to
tell the users to try again later.

Pin the dependency python-keycloak because they have some current
issues which would require changes to our code if fixed.  See
github.com/marcospereirampj/python-keycloak, issues 90, 95 and 99.

+++ app/auth.py
@@ -31,6 +31,13 @@ class KeycloakAdmin(KeycloakAdmin_):
         self.connection.add_param_headers('Authorization', 'Bearer ' + self.token.get('access_token'))
 
 
+class AuthError(Exception):
+    """Error used to refuse changes to users with Keycloak accounts when a
+    Keycloak server is not configured.
+    """
+    pass
+
+
 class AuthProvider:
     def __init__(self, app=None):
         if app is not None:
@@ -89,13 +96,12 @@ class AuthProvider:
                                                    'emailVerified': verified_email,
                                                    'credentials': [{'value': password,
                                                                     'type': 'password'}]})
-            # email = ''  # Store emails in local db, or not?
             password = ''
             crypto = UserCrypto.REMOTE
         user = User.create(uid=uid, name=name, crypto=crypto, password=password,
                            email=email, joindate=datetime.utcnow())
         self.set_user_auth_source(user, auth_source)
-        self.set_email_verified(user, verified_email)
+        self._set_email_verified(user, verified_email)
         return user
 
     def get_user_auth_source(self, user):
@@ -116,46 +122,71 @@ class AuthProvider:
         except UserMetadata.DoesNotExist:
             UserMetadata.create(uid=user.uid, key="auth_source", value=value)
 
-    def change_password(self, user, old_password, new_password):
-        if not self.validate_password(user, old_password):
-            return False
-        if self.provider == 'LOCAL':
-            user.crypto = UserCrypto.BCRYPT
-            user.password = bcrypt.hashpw(new_password.encode('utf-8'), bcrypt.gensalt())
-        elif self.provider == 'KEYCLOAK':
-            # validate_password promotes LOCAL users to KEYCLOAK ones.
-            self.keycloak_admin.update_user(user_id=user.uid,
-                                            payload={'credentials':
-                                                     [{'value': new_password,
-                                                       'type': 'password'}]})
-        # Invalidate other existing login sessions.
-        user.resets += 1
-        user.save()
-        theuser = misc.load_user(user.uid)
-        login_user(theuser, remember=session.get("remember_me", False))
+    def get_user_remote_uid(self, user):
+        try:
+            umd = UserMetadata.get((UserMetadata.uid == user.uid) &
+                                   (UserMetadata.key == "remote_uid"))
+            return umd.value
+        except UserMetadata.DoesNotExist:
+            return user.uid
 
-    def reset_password(self, user, new_password):
-        if self.provider == 'LOCAL':
-            user.crypto = UserCrypto.BCRYPT
-            user.password = bcrypt.hashpw(new_password.encode('utf-8'), bcrypt.gensalt())
-            self.set_user_auth_source(user, UserAuthSource.LOCAL)
-        elif self.provider == 'KEYCLOAK':
-            if self.get_user_auth_source(user) == UserAuthSource.LOCAL:
-                self.keycloak_admin.create_user({'id': user.uid,
-                                                 'email': user.email,
-                                                 'username': user.name,
-                                                 'enabled': True,
-                                                 'emailVerified': self.is_email_verified(user),
-                                                 'credentials': [{'value': new_password,
-                                                                  'type': 'password'}]})
-                self.set_user_auth_source(user, UserAuthSource.KEYCLOAK)
-                user.crypto = UserCrypto.REMOTE
-                user.password = ''
-            else:
-                self.keycloak_admin.update_user(user_id=user.uid,
+    def set_user_remote_uid(self, user, remote_uid):
+        try:
+            umd = UserMetadata.get((UserMetadata.uid == user.uid) &
+                                   (UserMetadata.key == "remote_uid"))
+            umd.value = remote_uid
+            umd.save()
+        except UserMetadata.DoesNotExist:
+            UserMetadata.create(uid=user.uid, key="remote_uid", value=remote_uid)
+        pass
+
+    def change_password(self, user, old_password, new_password):
+        if self.validate_password(user, old_password):
+            auth_source = self.get_user_auth_source(user)
+            if self.provider == 'LOCAL' and auth_source == UserAuthSource.LOCAL:
+                user.crypto = UserCrypto.BCRYPT
+                user.password = bcrypt.hashpw(new_password.encode('utf-8'),
+                                              bcrypt.gensalt())
+            elif self.provider == 'KEYCLOAK' and auth_source == UserAuthSource.KEYCLOAK:
+                self.keycloak_admin.update_user(user_id=self.get_user_remote_uid(user),
                                                 payload={'credentials':
                                                          [{'value': new_password,
                                                            'type': 'password'}]})
+            else:
+                raise AuthError
+            # Invalidate other existing login sessions.
+            user.resets += 1
+            user.save()
+            theuser = misc.load_user(user.uid)
+            login_user(theuser, remember=session.get("remember_me", False))
+
+    def reset_password(self, user, new_password):
+        auth_source = self.get_user_auth_source(user)
+        if self.provider == 'LOCAL':
+            if auth_source == UserAuthSource.LOCAL:
+                user.crypto = UserCrypto.BCRYPT
+                user.password = bcrypt.hashpw(new_password.encode('utf-8'), bcrypt.gensalt())
+            else:
+                raise AuthError
+        elif self.provider == 'KEYCLOAK':
+            if auth_source == UserAuthSource.LOCAL:
+                kuid = self.keycloak_admin.create_user({'email': user.email,
+                                                        'username': user.name,
+                                                        'enabled': True,
+                                                        'emailVerified': self.is_email_verified(user),
+                                                        'credentials': [{'value': new_password,
+                                                                         'type': 'password'}]})
+                self.set_user_remote_uid(user, kuid)
+                self.set_user_auth_source(user, UserAuthSource.KEYCLOAK)
+                user.crypto = UserCrypto.REMOTE
+                user.password = ''
+            elif auth_source == UserAuthSource.KEYCLOAK:
+                self.keycloak_admin.update_user(user_id=self.get_user_remote_uid(user),
+                                                payload={'credentials':
+                                                         [{'value': new_password,
+                                                           'type': 'password'}]})
+            else:
+                raise AuthError
         user.resets += 1
         user.save()
 
@@ -185,18 +216,29 @@ class AuthProvider:
             UserMetadata.create(uid=user.uid, key="pending_email", value=email)
 
     def confirm_pending_email(self, user, email):
+        umd = None
         try:
             umd = UserMetadata.get((UserMetadata.uid == user.uid) &
                                    (UserMetadata.key == "pending_email"))
-            umd.delete_instance()
         except UserMetadata.DoesNotExist:
             pass
-        if self.get_user_auth_source(user) == UserAuthSource.KEYCLOAK:
-            self.keycloak_admin.update_user(user_id=user.uid,
-                                            payload={'email': email})
+        auth_source = self.get_user_auth_source(user)
+        if self.provider == 'LOCAL':
+            if auth_source != UserAuthSource.LOCAL:
+                raise AuthError
+        elif self.provider == 'KEYCLOAK':
+            if auth_source == UserAuthSource.KEYCLOAK:
+                self.keycloak_admin.update_user(user_id=self.get_user_remote_uid(user),
+                                                payload={'email': email,
+                                                         'emailVerified': True})
+            else:
+                raise AuthError
+
+        if umd is not None:
+            umd.delete_instance()
         user.email = email
         user.save()
-        self.set_email_verified(user)
+        self._set_email_verified(user)
 
     def is_email_verified(self, user):
         try:
@@ -207,9 +249,21 @@ class AuthProvider:
             return False
 
     def set_email_verified(self, user, value=True):
-        if self.get_user_auth_source(user) == UserAuthSource.KEYCLOAK:
-            self.keycloak_admin.update_user(user_id=user.uid,
-                                            payload={'emailVerified': value})
+        """Set both the UserMetadata and remote (if any) email_verified flags."""
+        auth_source = self.get_user_auth_source(user)
+        if self.provider == 'LOCAL':
+            if auth_source != UserAuthSource.LOCAL:
+                raise AuthError
+        elif self.provider == 'KEYCLOAK':
+            if auth_source == UserAuthSource.KEYCLOAK:
+                self.keycloak_admin.update_user(user_id=self.get_user_remote_uid(user),
+                                                payload={'emailVerified': value})
+            else:
+                raise AuthError
+        self._set_email_verified(user, value)
+
+    def _set_email_verified(self, user, value=True):
+        """Set the UserMetadata email_verified flag. """
         value = '1' if value else '0'
         try:
             umd = UserMetadata.get((UserMetadata.uid == user.uid) &
@@ -224,13 +278,13 @@ class AuthProvider:
             thash = bcrypt.hashpw(password.encode('utf-8'), user.password.encode('utf-8'))
             if thash == user.password.encode('utf-8'):
                 if self.provider == 'KEYCLOAK':
-                    self.keycloak_admin.create_user({'id': user.uid,
-                                                     'email': user.email,
-                                                     'username': user.name,
-                                                     'enabled': True,
-                                                     'emailVerified': self.is_email_verified(user),
-                                                     'credentials': [{'value': password,
-                                                                      'type': 'password'}]})
+                    kuid = self.keycloak_admin.create_user({'email': user.email,
+                                                            'username': user.name,
+                                                            'enabled': True,
+                                                            'emailVerified': self.is_email_verified(user),
+                                                            'credentials': [{'value': password,
+                                                                             'type': 'password'}]})
+                    self.set_user_remote_uid(user, kuid)
                     self.set_user_auth_source(user, UserAuthSource.KEYCLOAK)
                     user.crypto = UserCrypto.REMOTE
                     user.password = ''
@@ -265,7 +319,8 @@ class AuthProvider:
         if user.crypto == UserCrypto.REMOTE:
             if (self.get_user_auth_source(user) == UserAuthSource.KEYCLOAK and
                     self.provider == 'KEYCLOAK'):
-                self.keycloak_admin.update_user(user_id=user.uid, payload=payload)
+                self.keycloak_admin.update_user(user_id=self.get_user_remote_uid(user),
+                                                payload=payload)
 
         user.status = new_status
         user.resets += 1  # Make them log in again.
@@ -275,7 +330,7 @@ class AuthProvider:
         # Used by automatic tests to clean up test realm on server.
         # You should probably be using mark_user_deleted.
         if user.crypto == UserCrypto.REMOTE:
-            self.keycloak_admin.delete_user(user.uid)
+            self.keycloak_admin.delete_user(self.get_user_remote_uid(user))
 
 
 auth_provider = AuthProvider()
